requirements.txt
Django==1.8
django-toolbelt==0.0.1
djangorestframework==3.1.0
static3==0.5.1
wsgiref==0.1.2